examples/02-pattern-examples/router-pattern/requirements.txt
# TFrameX Router Pattern Example Dependencies

# Core TFrameX framework
tframex>=0.1.3

# Required for environment configuration
python-dotenv>=1.0.0

# Optional: Enhanced logging
colorama>=0.4.6